Switch to port 12112 as default stripe-mock port for HTTPS
See stripe/stripe-mock#84 for context, but I'm going to switch to port
`12112` as the new default for stripe-mock running HTTPS. This will
allow us to have a setup where we can run API libraries that use HTTP
and those that use HTTPS with just a single instance of stripe-mock.

Depends on stripe/stripe-mock#84 being merged first, and a new
stripe-mock release being cut.